+++ cloudapi-20160714/src/Models/CreateStageModelRequest.php
@@ -1,0 +1,90 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\CloudAPI\V20160714\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class CreateStageModelRequest extends Model
+{
+    /**
+     * @var string
+     */
+    public $description;
+
+    /**
+     * @var string
+     */
+    public $securityToken;
+
+    /**
+     * @var string
+     */
+    public $stageAlias;
+
+    /**
+     * @var string
+     */
+    public $stageName;
+    protected $_name = [
+        'description' => 'Description',
+        'securityToken' => 'SecurityToken',
+        'stageAlias' => 'StageAlias',
+        'stageName' => 'StageName',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->description) {
+            $res['Description'] = $this->description;
+        }
+
+        if (null !== $this->securityToken) {
+            $res['SecurityToken'] = $this->securityToken;
+        }
+
+        if (null !== $this->stageAlias) {
+            $res['StageAlias'] = $this->stageAlias;
+        }
+
+        if (null !== $this->stageName) {
+            $res['StageName'] = $this->stageName;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Description'])) {
+            $model->description = $map['Description'];
+        }
+
+        if (isset($map['SecurityToken'])) {
+            $model->securityToken = $map['SecurityToken'];
+        }
+
+        if (isset($map['StageAlias'])) {
+            $model->stageAlias = $map['StageAlias'];
+        }
+
+        if (isset($map['StageName'])) {
+            $model->stageName = $map['StageName'];
+        }
+
+        return $model;
+    }
+}