+++ cloudapi-20160714/src/Models/CreateStageModelResponseBody.php
@@ -1,0 +1,62 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\CloudAPI\V20160714\Models;
+
+use AlibabaCloud\Dara\Model;
+
+class CreateStageModelResponseBody extends Model
+{
+    /**
+     * @var string
+     */
+    public $requestId;
+
+    /**
+     * @var string
+     */
+    public $stageModelId;
+    protected $_name = [
+        'requestId' => 'RequestId',
+        'stageModelId' => 'StageModelId',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+
+        if (null !== $this->stageModelId) {
+            $res['StageModelId'] = $this->stageModelId;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+
+        if (isset($map['StageModelId'])) {
+            $model->stageModelId = $map['StageModelId'];
+        }
+
+        return $model;
+    }
+}